+++ src/OpenLoco/Config.cpp
@@ -164,6 +164,8 @@ namespace OpenLoco::Config
             _new_config.language = config["language"].as<std::string>();
         if (config["breakdowns_disabled"])
             _new_config.breakdowns_disabled = config["breakdowns_disabled"].as<bool>();
+        if (config["trains_reverse_at_signals"])
+            _new_config.trains_reverse_at_signals = config["trains_reverse_at_signals"].as<bool>();
         if (config["cheats_menu_enabled"])
             _new_config.cheats_menu_enabled = config["cheats_menu_enabled"].as<bool>();
         if (config["companyAIDisabled"])
@@ -226,6 +228,7 @@ namespace OpenLoco::Config
         node["last_save_path"] = _new_config.last_save_path;
         node["language"] = _new_config.language;
         node["breakdowns_disabled"] = _new_config.breakdowns_disabled;
+        node["trains_reverse_at_signals"] = _new_config.trains_reverse_at_signals;
         node["cheats_menu_enabled"] = _new_config.cheats_menu_enabled;
         node["companyAIDisabled"] = _new_config.companyAIDisabled;
         node["scale_factor"] = _new_config.scale_factor;