+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 
 21.09.1 (2021-09-14)
 ------------------------------------------------------------------------
+- Feature: Add option to disable train reversing at signals
 - Fix: [#1161] Stuttering vehicle movement due to incorrect vehicle positions.
 - Fix: [#1162] Signal side placement is inverted.
 - Fix: [#1162] Single side signal removal incorrectly removing both signals.